Refactor GroupDateTime
